This is a creation of the UI,Buttons,and basic variables.
This commit contains data that has basic methods in  Clock model that
will be getters and setters.  The date model is an empty class for now
but will be the same as clock model.  The timeview contains all the
buttons tied to the UI for time purposes.  The command class will only
contain one method call execute.  Time change controller contais methods
that will get the current time for the system.